LPS-106140 Make sure the layout still exists as it might be a child layout of a parent layout that has already been deleted

+++ modules/apps/web-experience/export-import/export-import-service/src/main/java/com/liferay/exportimport/controller/LayoutImportController.java
@@ -316,6 +316,12 @@ public class LayoutImportController implements ImportController {
 			if (!sourceLayoutUuids.contains(layout.getUuid()) &&
 				!layoutPlids.containsValue(layout.getPlid())) {
 
+				layout = _layoutLocalService.fetchLayout(layout.getPlid());
+
+				if (layout == null) {
+					continue;
+				}
+
 				try {
 					_layoutLocalService.deleteLayout(
 						layout, false, serviceContext);
